Move tests to _testcapi/abstract
